requirements.txt
firebase-admin
google-cloud-firestore
google-auth
google-auth-oauthlib
google-auth-httplib2
streamlit
altair
pandas
numpy
toml
